fix failing test and typescript errors

+++ src/test/setup.ts
@@ -3,40 +3,42 @@ import mongoose from "mongoose";
 import jwtSign from "../helpers/token";
 
 declare global {
-      var customerSignIn: () => string
-      var agentSignIn: () => string
-      var adminSignIn: () => string
+  var customerSignIn: () => string;
+  var agentSignIn: () => string;
+  var adminSignIn: () => string;
 }
 let mongo: any;
 beforeAll(async () => {
   process.env.JWT_KEY = "adsfdg";
   process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
-  mongo = new MongoMemoryServer();
-  const mongoUri = await mongo.getUri();
 
-  await mongoose.connect(mongoUri, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  });
+  mongo = await MongoMemoryServer.create();
+  const mongoUri = mongo.getUri();
+
+  await mongoose.connect(mongoUri);
 });
 
 beforeEach(async () => {
-  const collections = await mongoose.connection.db.collections();
-
-  for (let collection of collections) {
-    await collection.deleteMany({});
+  if (!mongoose.connection.db) {
+    throw new Error("Database not connected");
   }
 
-  afterAll(async () => {
-    await mongo.stop();
-    await mongoose.connection.close();
-  });
+  const collections = await mongoose.connection.db.collections();
+
+  for (const collection of collections) {
+    await collection.deleteMany({});
+  }
+});
+
+afterAll(async () => {
+  await mongoose.connection.close();
+  await mongo.stop();
 });
 
 global.customerSignIn = () => {
   // Build a JWT payload.  { id, email and customer }
   const payload = {
-    id: mongoose.Types.ObjectId().toHexString(),
+    id: new mongoose.Types.ObjectId(),
     email: "test@test.com",
     role: "customer",
   };
@@ -50,7 +52,7 @@ global.customerSignIn = () => {
 global.agentSignIn = () => {
   // Build a JWT payload.  { id, email, and role }
   const payload = {
-    id: mongoose.Types.ObjectId().toHexString(),
+    id: new mongoose.Types.ObjectId(),
     email: "agent@gmail.com",
     role: "agent",
   };
@@ -64,7 +66,7 @@ global.agentSignIn = () => {
 global.adminSignIn = () => {
   // Build a JWT payload.  { id, email, and role }
   const payload = {
-    id: mongoose.Types.ObjectId().toHexString(),
+    id: new mongoose.Types.ObjectId(),
     email: "admin@gmail.com",
     role: "admin",
   };
